+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- limit collected type attributes by declared "type"
+
+### Fixed
+- not declare any "type" for `Object.class` by default
 
 ## [4.2.0] - 2020-02-27
 ### Added